+++ PP_OK_GC/PP_OK_GC/AntSLF.h
@@ -14,7 +14,7 @@ class AntSLF {
 	vector<int> c_min; //c_min[x] - minimalny kolor jaki mo¿e byæ przyporz¹dkowany wierzcho³kowi x
 	vector<int> dsat; //dsat[x] - stopieñ nasycenia wierzcho³ka x
 	vector<int> verticesWithoutColor; //zbiór wierzcho³ków nie pokolorowanych
-	vector<vector<int>> coloringQuality; //iloœæ feromoonu
+	vector<vector<float>> coloringQuality; //iloœæ feromoonu
 
 	int znajdzWierzcholekONajwiekszymStopniu(vector<int>);
 	void arraysInicjalization(int);
@@ -28,7 +28,7 @@ class AntSLF {
 	float chooseVertice_calculateT2(int, int);
 
 public:
-	AntSLF(vector<vector<int>>, int, vector<vector<int>>);
+	AntSLF(vector<vector<int>>, int, vector<vector<float>>);
 	vector<int> listSetColorVertices;
 	int numberOfColors; //iloœæ u¿ytych kolorów
 	vector<vector<int>> subsetsOfVertices; //podzbiowy wierzcho³ków
